scripts/tenant-discovery/requirements.txt
requests>=2.31.0,<3.0.0
pyzabbix>=1.3.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
